purego: use go:nosplit and map to proper implementation (#6)
This commit also adds windows support to purego
